Add parenting system, enable, visible and hooked booleans

+++ include/entities/ui/Button.h
@@ -36,8 +36,8 @@ class Button : public UIElement {
    * @param z Z position of element (for sorting)
    */
   explicit Button(Scene& scene,
-                  const int x = 0,
-                  const int y = 0,
+                  const float x = 0,
+                  const float y = 0,
                   const int z = 0);
 
   /**
